src/app/requirements.txt
boto3>=1.26.79
PyYAML==6.0.1
pillow==10.2.0